Add files via upload
A new geothermal heat flow model for the Arctic region (from 51°N to 90°N). This is a part of the dataset for the paper "A set of geophysical fields for modelling of the lithosphere structure and dynamics in the Russian Arctic zone". All the data will soon be available ath the World Data Center for Solid Earth Physics website.