Add files via upload
Moho depth data for an Eastern Eurasia Arctic area (60–75°N, 110–190°E) This is a part of the dataset for the paper "A set of geophysical fields for modelling of the lithosphere structure and dynamics in the Russian Arctic zone". All the data will soon be available ath the World Data Center for Solid Earth Physics website.