Add files via upload
Velocity anomaly depth slices from SL2013sv, SEMum2, SMEAN2 and S40RTS seismic tomography models from the surface (0 km, except for SL2013sv model) to 650 km (an average depth of the boundary between the upper and lower mantle). The depth step is 50 km. The spatial coverage is 60°N to 90N and 0° to 360° longitude.