Add files via upload
The dataset includes two sedimentary cover thickness models obtained from decompensation gravity anomalies data (Figure 4 displays the first of these two models, the second one differs in its initial conditions [43]), density correction map, and the map of the corrected density, averaged vertically. This is a complex of results of the analysis of structure of sedimentary basins in the Eastern Asia Arctic zone by employing the mentioned approach based on decompensative gravity anomalies. The data are freely accessible. The study area spans from 135° to 190°E and 65° to 74°N,m although the initial area was larger and included the 5600x5400 km in the orthographic projection with a center point 162.5°E, 72.5°N. This was done for more convenient calculation. The structure of the grids for sedimentary models is similar to the structure of the grids for gravity fields.